requirements.txt
Flask
sqlite3
google-cloud-vertexai
google-api-core 